requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
filelock==3.16.1
fsspec==2024.9.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
pillow==10.4.0
psycopg==3.2.3
python-dotenv==1.0.1
sqlparse==0.5.1
sympy==1.13.3
torch==2.4.1
torch-vision==0.1.6.dev0
torchaudio==2.4.1
torchvision==0.19.1
triton==3.0.0
typing_extensions==4.12.2
